+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Remove fire from Lara in DOZY.
 - Make footprints tilt with the floor below them.
+- Make PSX shadow tilt with the floor below it.
 
 ### tomb3 bug fixes